[frontend] Move packed wire population functions to util
