+++ crates/frontend/src/util.rs
@@ -1,6 +1,11 @@
 use std::{fmt, time::Instant};
 
-use crate::compiler::circuit::Circuit;
+use binius_core::Word;
+
+use crate::compiler::{
+	Wire,
+	circuit::{Circuit, WitnessFiller},
+};
 
 /// Various stats of a circuit that affect the prover performance.
 pub struct CircuitStat {
@@ -59,4 +64,84 @@ fn time_witness_population(circuit: &Circuit) {
 	let _ = circuit.populate_wire_witness(&mut w);
 	let elapsed = start.elapsed();
 	println!("fill_witness took {} microseconds", elapsed.as_micros());
+}
+
+/// Populate the given wires with packed 64-bit words from a byte slice.
+///
+/// The packing strategy is determined by the provided `pack` function, which takes
+/// an up-to-8-byte slice and returns a `u64` value.
+///
+/// If `bytes` is not a multiple of 8, the last word is zero-padded.
+///
+/// If there are more wires than needed to hold all bytes, the remaining wires
+/// are filled with `Word::ZERO`.
+///
+/// # Panics
+/// * If bytes.len() exceeds wires.len() * 8
+fn pack_bytes_into_wires(
+	w: &mut WitnessFiller,
+	wires: &[Wire],
+	bytes: &[u8],
+	pack: fn(&[u8]) -> u64,
+) {
+	let max_value_size = wires.len() * 8;
+	assert!(
+		bytes.len() <= max_value_size,
+		"bytes length {} exceeds maximum {}",
+		bytes.len(),
+		max_value_size
+	);
+
+	// Pack bytes into words
+	for (i, chunk) in bytes.chunks(8).enumerate() {
+		if i < wires.len() {
+			let word = pack(chunk);
+			w[wires[i]] = Word(word);
+		}
+	}
+
+	// Zero out remaining words
+	for i in bytes.len().div_ceil(8)..wires.len() {
+		w[wires[i]] = Word::ZERO;
+	}
+}
+
+/// Populate the given wires from bytes using little-endian packed 64-bit words.
+///
+/// If `bytes` is not a multiple of 8, the last word is zero-padded.
+///
+/// If there are more wires than needed to hold all bytes, the remaining wires
+/// are filled with `Word::ZERO`.
+///
+/// # Panics
+/// * If bytes.len() exceeds wires.len() * 8
+pub(crate) fn pack_bytes_into_wires_le(w: &mut WitnessFiller, wires: &[Wire], bytes: &[u8]) {
+	fn pack_le(chunk: &[u8]) -> u64 {
+		let mut word = 0u64;
+		for (j, &byte) in chunk.iter().enumerate() {
+			word |= (byte as u64) << (j * 8)
+		}
+		word
+	}
+	pack_bytes_into_wires(w, wires, bytes, pack_le);
+}
+
+/// Populate the given wires from bytes using big-endian packed 64-bit words.
+///
+/// If `bytes` is not a multiple of 8, the last word is zero-padded.
+///
+/// If there are more wires than needed to hold all bytes, the remaining wires
+/// are filled with `Word::ZERO`.
+///
+/// # Panics
+/// * If bytes.len() exceeds wires.len() * 8
+pub(crate) fn pack_bytes_into_wires_be(w: &mut WitnessFiller, wires: &[Wire], bytes: &[u8]) {
+	fn pack_be(chunk: &[u8]) -> u64 {
+		let mut word = 0u64;
+		for (j, &byte) in chunk.iter().enumerate() {
+			word |= (byte as u64) << ((7 - j) * 8)
+		}
+		word
+	}
+	pack_bytes_into_wires(w, wires, bytes, pack_be);
 }
